# build_T speedup using numpy threading 