--- a/environment.yaml
+++ b/environment.yaml
@@ -22,3 +22,4 @@ dependencies:
   - setuptools
   - setuptools_scm
   - sphinx
+  - threadpoolctl