@jburba 2025-02-14
>     1. Do we need to account for nq when calculating the time indices? E.g. should it be
>        `time_inds = slice( i_t*self.nbls*(self.nf+self.nq), (i_t + 1)*self.nbls*(self.nf+self.nq) )`

I don't think we need nq in the time indices here as the time indices are only used to index the visibility axis of Finv which has shape (nt*nf*nbls,).  The nq terms are only relevant in the seond axis of Fprime_Fz which has shape (nhpx*nf, nf*(nu_eor*nv_eor - 1), (nq + 1)*(nu_fg*nv_fg - 1) + nf*fit_for_monopole), i.e. the first axis of Fprime_Fz indexes the sky model and the second axis indexes the joint EoR and foreground model.  But, we don't need to worry about the second axis of Fprime_Fz here, so I think we can ignore nq altogether in the indices used to extract the subset arrays.  The test I ran also used nq=2, so the plot in my previous comment shows the two matrices agree when nq > 0.


 
>     2. I like your flattened residuals check! To be safe, could you run it for nq = 0, 1 and 2? The residuals test is useful generally, especially if we end up wanting to use a similar approach to speed-up any other matrices. What's your opinion on adding it as a unit test?

I can run it for a few nq values, but I don't know if that's necessary given my answer to your previous question.  Let me know if I'm wrong and/or missed something, though.  And, as I previously mentioned, the plot above compares T for nq=2, so the results agree for nq > 0.  I suspect they will agree for nq=0, too, but I'm happy to check that's the case if need be.

Adding it as a unit test is probably a good idea.  We've been talking about adding some tests for a while now.  I think it would be pretty easy to cook this up as a test with some small model parameters to make things faster.