@PSims 2025-02-15
That sounds reasonable to me, in which case I think this is good to merge!

Converting your check to a unit test that can be run quickly on a small test-case also sounds great.